move AggregationStore inside Table class and handle both static and aggregations cache in one place
