+++ tests/test_flow.py
@@ -460,8 +460,8 @@ def test_map_with_state_flow():
 
 def test_map_with_cache_state_flow():
     table_object = Table("table", NoopDriver())
-    table_object._cache['tal'] = {'color': 'blue'}
-    table_object._cache['dina'] = {'color': 'red'}
+    table_object.set_static_attrs('tal', {'color': 'blue'})
+    table_object.set_static_attrs('dina', {'color': 'red'})
 
     def enrich(event, state):
         event['color'] = state['color']
@@ -495,7 +495,9 @@ def test_map_with_cache_state_flow():
     expected_cache = {'tal': {'color': 'blue', 'counter': 6}, 'dina': {'color': 'red', 'counter': 4}}
 
     assert termination_result == expected
-    assert table_object._cache == expected_cache
+    assert len(table_object._attrs_cache) == len(expected_cache)
+    assert table_object.get_static_attrs('tal') == expected_cache['tal']
+    assert table_object.get_static_attrs('dina') == expected_cache['dina']
 
 
 def test_map_with_empty_cache_state_flow():
@@ -533,8 +535,9 @@ def test_map_with_empty_cache_state_flow():
                 {'col1': 8, 'diff_from_first': 7},
                 {'col1': 9, 'diff_from_first': 9}]
     expected_cache = {'dina': {'first_value': 0, 'counter': 4}, 'tal': {'first_value': 1, 'counter': 6}}
-    assert termination_result == expected
-    assert table_object._cache == expected_cache
+    assert len(table_object._attrs_cache) == len(expected_cache)
+    assert table_object.get_static_attrs('tal') == expected_cache['tal']
+    assert table_object.get_static_attrs('dina') == expected_cache['dina']
 
 
 def test_awaitable_result():
@@ -1529,8 +1532,8 @@ def test_write_to_parquet_with_inference(tmpdir):
 
 def test_join_by_key():
     table = Table('test', NoopDriver())
-    table.update_key(9, {'age': 1, 'color': 'blue9'})
-    table.update_key(7, {'age': 3, 'color': 'blue7'})
+    table.update_static_attrs(9, {'age': 1, 'color': 'blue9'})
+    table.update_static_attrs(7, {'age': 3, 'color': 'blue7'})
 
     controller = build_flow([
         Source(),
@@ -1549,8 +1552,8 @@ def test_join_by_key():
 
 def test_join_by_string_key():
     table = Table('test', NoopDriver())
-    table.update_key(9, {'age': 1, 'color': 'blue9'})
-    table.update_key(7, {'age': 3, 'color': 'blue7'})
+    table.update_static_attrs(9, {'age': 1, 'color': 'blue9'})
+    table.update_static_attrs(7, {'age': 3, 'color': 'blue7'})
 
     controller = build_flow([
         Source(),
